server/monolith/requirements.txt
Django
Celery
djangorestframework
djangorestframework-simplejwt
django-filter
python-dotenv
psycopg2-binary==2.8.6
Pillow
